requirements.txt
appdirs==1.4.4
asgiref==3.2.7
backcall==0.2.0
decorator==4.4.2
distlib==0.3.0
Django==3.0.7
djangorestframework==3.11.0
filelock==3.0.12
flake8==3.8.3
gunicorn==20.0.4
importlib-metadata==1.6.1
importlib-resources==2.0.1
ipdb==0.13.2
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
mccabe==0.6.1
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkgconfig==1.5.1
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
virtualenv==20.0.23
virtualfish==2.3.0
wcwidth==0.2.4
zipp==3.1.0
